move hexstring in one location
(and do not produce 'bytes', since callers expect ASCII)
