+++ lib/exabgp/bgp/message/keepalive.py
@@ -9,9 +9,8 @@ Copyright (c) 2009-2015 Exa Networks. All rights reserved.
 from exabgp.bgp.message.message import Message
 from exabgp.bgp.message.notification import Notify
 
-from exabgp.reactor.api.options import hexstring
-
 from exabgp.util import character
+from exabgp.util import hexstring
 
 # =================================================================== KeepAlive
 #
