+++ lib/exabgp/reactor/api/options.py
@@ -22,10 +22,3 @@ class APIOptions (dict):
 
 	def __missing__ (self, key):
 		return False
-
-
-def hexstring (value):
-	def spaced (value):
-		for v in value:
-			yield '%02X' % ord(v)
-	return b''.join(spaced(value))
