+++ lib/exabgp/reactor/api/response/json.py
@@ -12,7 +12,7 @@ import socket
 import sys
 import time
 
-from exabgp.reactor.api.options import hexstring
+from exabgp.util import hexstring
 
 from exabgp.bgp.message import Message
 from exabgp.bgp.message import IN
