+++ lib/exabgp/reactor/api/transcoder.py
@@ -6,6 +6,7 @@ import json
 
 from exabgp.util import character
 from exabgp.util import concat_bytes
+from exabgp.util import hexstring
 
 from exabgp.bgp.message import Message
 from exabgp.bgp.message import Open
